+++ jest.config.js
@@ -10,5 +10,7 @@ module.exports = {
     },
   },
   transformIgnorePatterns: ['<rootDir>/node_modules/'],
-  modulePathIgnorePatterns: ["<rootDir>/tests/kms"]
+  modulePathIgnorePatterns: [
+    "<rootDir>/tests/kms",
+  ]
 };